backend/requirements.txt
pillow
cloudinary
fastapi
passlib
pydantic
pyjwt
pymongo
python-decouple
uvicorn
bcrypt
python-multipart
ffmpeg-python
cryptography
